CNV-87886: remove ListPageFilter and related components
