+++ src/utils/hooks/useKubevirtDataViewFilters/filters/getLabelFilter.ts
@@ -1,7 +1,7 @@
 import { TFunction } from 'i18next';
 
-import { STATIC_SEARCH_FILTERS } from '@kubevirt-utils/components/ListPageFilter/constants';
-import { getLabelsAsString } from '@kubevirt-utils/components/ListPageFilter/utils';
+import { STATIC_SEARCH_FILTERS } from '@kubevirt-utils/components/KubevirtFilterToolbar/constants';
+import { getLabelsAsString } from '@kubevirt-utils/utils/labelUtils';
 import {
   FilterableObject,
   KubevirtFilter,
