+++ src/utils/hooks/useKubevirtDataViewFilters/filters/getNameFilter.ts
@@ -1,12 +1,12 @@
 import { TFunction } from 'i18next';
 
-import { STATIC_SEARCH_FILTERS } from '@kubevirt-utils/components/ListPageFilter/constants';
-import { fuzzyCaseInsensitive } from '@kubevirt-utils/components/ListPageFilter/utils';
+import { STATIC_SEARCH_FILTERS } from '@kubevirt-utils/components/KubevirtFilterToolbar/constants';
 import {
   FilterableObject,
   KubevirtFilter,
 } from '@kubevirt-utils/hooks/useKubevirtDataViewFilters/types';
 import { getName } from '@kubevirt-utils/resources/shared';
+import { fuzzyCaseInsensitive } from '@kubevirt-utils/utils/utils';
 
 export const getNameFilter = (t: TFunction): KubevirtFilter<FilterableObject> => ({
   categoryLabel: t('Name'),
