--- src/utils/hooks/useProjectFilter.ts
@@ -1,58 +1,0 @@
-import { useMemo } from 'react';
-
-import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
-import useProjects from '@kubevirt-utils/hooks/useProjects';
-import { getName, getNamespace } from '@kubevirt-utils/resources/shared';
-import { PROJECT_LIST_FILTER_TYPE } from '@kubevirt-utils/utils/constants';
-import { isEmpty, universalComparator } from '@kubevirt-utils/utils/utils';
-import useMulticlusterNamespaces from '@multicluster/hooks/useMulticlusterNamespaces';
-import useIsACMPage from '@multicluster/useIsACMPage';
-import { type RowFilter } from '@openshift-console/dynamic-plugin-sdk';
-
-import useListClusters from './useListClusters';
-
-export const useProjectFilter = <R extends K8sResourceCommon>(): RowFilter<R> => {
-  const { t } = useKubevirtTranslation();
-  const allClustersSelected = useListClusters();
-  const isACMPage = useIsACMPage();
-
-  const [projects] = useProjects();
-  const { allNamespaces: multiclusterNamespaces } = useMulticlusterNamespaces(allClustersSelected);
-
-  const multiclusterNamespacesNames = useMemo(
-    () =>
-      multiclusterNamespaces
-        ?.reduce((acc, project) => {
-          const projectName = getName(project);
-          if (!acc.includes(projectName)) {
-            acc.push(projectName);
-          }
-          return acc;
-        }, [] as string[])
-        ?.sort((a, b) => universalComparator(a, b)) || [],
-    [multiclusterNamespaces],
-  );
-
-  const projectNames = useMemo(
-    () => (isACMPage ? multiclusterNamespacesNames : (projects ?? [])),
-    [isACMPage, multiclusterNamespacesNames, projects],
-  );
-
-  return {
-    filter: (input, obj): boolean => {
-      if (isEmpty(input.selected)) {
-        return true;
-      }
-
-      const namespace = getNamespace(obj);
-      return !!namespace && input.selected.includes(namespace);
-    },
-    filterGroupName: t('Project'),
-    isMatch: () => true,
-    items: projectNames.map((project) => ({
-      id: project,
-      title: project,
-    })),
-    type: PROJECT_LIST_FILTER_TYPE,
-  };
-};
