+++ src/views/search/components/AdvancedSearchModal/formFields/LabelsField.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useMemo } from 'react';
 
 import { V1VirtualMachine } from '@kubevirt-ui-ext/kubevirt-api/kubevirt';
-import { labelParser } from '@kubevirt-utils/components/ListPageFilter/utils';
+import { labelParser } from '@kubevirt-utils/utils/labelUtils';
 import MultiSelectTypeahead from '@kubevirt-utils/components/MultiSelectTypeahead/MultiSelectTypeahead';
 import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
 import { FormGroup } from '@patternfly/react-core';
